[DEV] Add the device steering screen

+++ data/src/main/java/me/pgloaguen/data/mapper/DeviceMapper.kt
@@ -38,7 +38,7 @@ class DeviceMapper {
         Device.Heater(
             item.id!!,
             item.deviceName!!,
-            item.temperature!!,
+            item.temperature!!.toFloat(),
             transformMode(item.mode),
         )
